Remove redundant menu icon which did nothing
In narrow viewports, a menu button was visible which had no behaviour associated with it. I believe this was intended to show the menu which is hidden on narrow viewports. We don't really use the menu at the moment as we don't have other pages. This may be more important if we introduce more pages into the blog, for example a video page. I've created a `menu-button-present-but-inactive` branch with these pages and issue #2 representing the fact that in narrow screens there will be no way to see the menu.